Don't crash when repo isn't visible
If a repo has been tagged with `govuk`, but isn't in the main GOV.UK
team, the authenticating user (govuk-ci, who created `GITHUB_TOKEN`)
will not have permission to update the repo. This will cause a
`Octokit::NotFound` error (because the user can't even see the repo).

In this case we can just skip the repo updating.