Force a new review if a PR is changed after a review
This commit turns on the GitHub option that will automatically dismiss a review if the PR is changed after the review has been submitted. This forces a new review before the PR can be merged.

We're testing this setting for a week and will determine the impact before deciding whether to keep it or revert it.