Allow squash merging on smart answers
It can sometimes be useful to squash all commits before merging a
branch. Currently this is disallowed on all our repos.

A good example of where this is useful is when dealing with content
requests. Often they will be using the GitHub website interface meaning
that PRs can end up with many commits, each changing a small part of the
content and even reverting older changes during the content 2i process.
Ultimately, the content change is more useful as a single commit so
often developers will squash all their commits before merging. This
takes up time when we could use the button in GitHub.