Restrict PR merge for Smart Answers
Previously we agreed that on a beta Continuous Deployment involving
this repo, with the caveat that any changes should only be mergeable
by people who would otherwise be manually deploying them. This adds
a merge restriction to the existing 'master' branch protections for
Smart Answers, with scope to add more repos in future.