Restrict merge access to most GOV.UK repos
This restriction is necessary for apps with Continuous Deployment
enabled i.e. where "merge" is the same as "deploy". Currently, the
set of developers without production access may include people without
security clearance, for whom we have to assume a worst case scenario.

As part of moving to Continuous Deployment, we will incrementally
override each repo to add this restriction. Rather than adding lots
of overrides, we should try to make a clear policy decision about
whether non-production developers should be able to merge PRs.

Exceptions to this:

- govuk-puppet - some changes only apply to Integration, such as
initial PRs for someone joining GOV.UK [1]

Reasons for this change:

- We avoid building up a set of overrides for Continuous Deployment.

- We more clearly align deploying and change with merging it (since
the person merging it is always able to).

- People deploying changes can have more visibility of them, having
merged the changes they are deploying themselves.

- A higher barrier to merge for new joiners should promote the good
habit of doing branch deploys for any manual testing.

Reasons against this change:

- Some loss of satisfaction of merging your own PR, even though you
couldn't actually deploy it to Production.

- Some additional workload on developers with production access, who
will now be asked to merge PRs, as well as deploy them.

[1]: https://docs.publishing.service.gov.uk/manual/get-started.html#6-get-ssh-access-to-integration