Change GitHub branch protection enforce admins to false
There's currently issues with the Publishing E2E Tests Jenkins job
updating the GitHub commit status, which blocks merging Pull Requests
because this is a required check.

To make it easier for people to check the Publishing E2E Tests passed,
and then merge regardless of the pending check, change enforce admins
to false. This should be reverted once the Jenkins -> GitHub status
reporting is fixed.