Add create single project, refactor create all
When we went to create the account-api project with create all, it began
creating sentry projects for many different non-account based teams.
This was worrying and suggested that our apps were slightly out of sync.

Until we sort out why, it makes sense for us to be able to create a
singular app that we specify.

This creates a separate rake task that allows for that, whilst re-using
much of the code and keeping things dry.