WIP: try out swapping out Octokit::Client mock
Arguably leads to cleaner code, as we don't care about the internal
implementation of the Octokit::Client, whether it makes 1 request
or 100 requests. Feels a bit gross to be stubbing a network
request made by a third party library. Providing our own mock
object that fulfils the same code-level API is potentially cleaner,
but as you can see here, also requires re-architecting things a
bit so as not to break the existing tests.

Something to consider exploring.