Protect gh-pages branch
This is a post-incident action:
https://docs.google.com/document/d/1SO8i3Mi4MLOKD-_Hf1kyXNPduOv1KNhihjZ4YXf1x20/edit

By protecting this branch, engineers will be forced to go into
the repo settings and explicitly remove the branch protection in
order to delete the gh-pages branch. This should avoid any
accidental deletion.

This added code has been manually tested against govuk-developer-docs
(where it successfully set up gh-pages branch protection after
I manually deleted the protection), and against another GOV.UK
repo which did _not_ have a gh-pages branch (and which failed with
"Could not find alphagov/repo. Possibly the govuk-ci user doesn't
have admin access to this repo.". Evidently the call to
`client.protect_branch` against the gh-pages branch, in a repo
which does not have said branch, would raise an `Octokit::NotFound`
exception, causing the rest of the configuration to be cancelled.
I've therefore added a handler for the exception, within the method.

The unit test for the exception-handling is not very well written
(relying on tightly coupled implementation details) but as our
GitHub configuration is destined to be moved to Terraform in the
near future, a suboptimal test seems fine for now.