Require up-to-date branch on merge for Terraform GitOps repos.
Generally this is best avoided because it's inconvenient, but for
continuously-deployed / GitOps-style Terraform repos it's the lesser of
evils because otherwise the pre-merge diff can be confusing if the
author forgets to update/rebase their PR branch.

We've had this set via the UI before, but it was far from obvious that
govuk-saas-config was unsetting it again even though the option wasn't
specified either way for these repos in its config here.