Enable need_production_access_to_merge by default.
This flag was `true` on almost every repo and is a somewhat important
security control, so make it the default.

Explicitly set it to false on the few repos (which are all docs repos)
where it wasn't already enabled.

This also enables the flag (by default) on a few repos which should have
had it but didn't.