Move all repos under a `repos` property
In the next commit we'll define some aliases to make managing the
'standard' CI status checks more consistent (a la
govuk-user-reviewer), and also to highlight where repos differ
from that standard.

The aliases will be defined under a new top-level property in this
file, so we need to namespace our array of repos to avoid the
aliases being treated as repos themselves.