Allow new `standard_contexts` property
This will allow engineers to mix and match both `standard_contexts`
and `additional_contexts` interchangeably. This paves the way for
us referencing aliases with the `standard_contexts` property, and
providing additional contexts with the `additional_contexts`
property, as YAML does not provide a native way of merging
arbitrary values and aliased values in the same property (it
only lets you overwrite).