Review & add status checks that had not been set as required
This double-checks the names of, and then marks as required, any
status checks that probably ought to be required (following other
precedents like linters, which are reliable and which we tend to
mark as required). The ERB and Prettier linters should be treated
the same as the Rubocop, Standard (JS) and SCSS linters.

Pact tests, on the other hand, are less black and white in nature.
Sometimes a Pact test will fail expectedly, and it's only after
merging a dependent PR that re-running both provider/consumer
Pact test suites will pass. So it makes sense for these to be
omitted from the required status checks.